requirements.txt
customtkinter==5.2.2
PyMuPDF==1.23.26
python-docx==1.1.0
pyinstaller==6.4.0
